workflow changes and removed old notifications api references
