+++ .github/workflows/release-branch.yml
@@ -81,9 +81,9 @@ jobs:
       - name: Configure Env Vars (secrets)
         run: scripts/environment-variables-secrets.sh
         env:
+          CAMPD_API_KEY: ${{ secrets.API_KEY }}
           CAMPD_CLIENT_ID: ${{ secrets.CLIENT_ID }}
           CAMPD_CLIENT_SECRET: ${{ secrets.CLIENT_SECRET }}
-          CAMPD_API_KEY: ${{ secrets.CAMPD_NON_PROD_API_KEY }}
       - name: Deployment
         run: devops/scripts/deploy.sh
 
@@ -122,9 +122,9 @@ jobs:
       - name: Configure Env Vars (secrets)
         run: scripts/environment-variables-secrets.sh
         env:
+          CAMPD_API_KEY: ${{ secrets.API_KEY }}
           CAMPD_CLIENT_ID: ${{ secrets.CLIENT_ID }}
           CAMPD_CLIENT_SECRET: ${{ secrets.CLIENT_SECRET }}
-          CAMPD_API_KEY: ${{ secrets.CAMPD_NON_PROD_API_KEY }}
       - name: Deployment
         run: devops/scripts/deploy.sh
 
@@ -163,8 +163,8 @@ jobs:
       - name: Configure Env Vars (secrets)
         run: scripts/environment-variables-secrets.sh
         env:
+          CAMPD_API_KEY: ${{ secrets.API_KEY }}
           CAMPD_CLIENT_ID: ${{ secrets.CLIENT_ID }}
           CAMPD_CLIENT_SECRET: ${{ secrets.CLIENT_SECRET }}
-          CAMPD_API_KEY: ${{ secrets.CAMPD_NON_PROD_API_KEY }}
       - name: Deployment
         run: devops/scripts/deploy.sh
